kadai auth crud

+++ app/Http/Controllers/TasksController.php
@@ -18,11 +18,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
         
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            
+            return view('tasks.index', $data);
+        }
+        
+        return view('welcome');
     }
 
     /**
@@ -51,13 +61,13 @@ class TasksController extends Controller
             'content' => 'required|max:255',
             'status' => 'required|max:10',
         ]);
+
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-        
-        return redirect('/');
+        return redirect('/tasks');
     }
 
     /**
@@ -68,11 +78,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        $user = \Auth::user();
         $task = Task::find($id);
+    
+        if ($user->id === $task->user_id) {
+            return view('tasks.show', [
+               'task' => $task, 
+            ]);
+        }
         
-        return view('tasks.show', [
-           'task' => $task, 
-        ]);
+        return redirect('/');
     }
 
     /**
@@ -105,6 +120,7 @@ class TasksController extends Controller
         ]);
         
         $task = Task::find($id);
+        $task = $request->user()->tasks()->find($id);
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -121,7 +137,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
+        
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
